add ( on_boarding ) : add page_select
