+++ lib/presentation/screens/on_boarding_v2/on_boarding_screen_v2.dart
@@ -1,3 +1,5 @@
+import 'package:finniu/presentation/providers/settings_provider.dart';
+import 'package:finniu/presentation/screens/on_boarding_v2/widgets/page_select.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -30,11 +32,45 @@ class StackOnBoardingState extends ConsumerState<StackOnBoarding> {
       height: MediaQuery.of(context).size.height,
       child: Stack(
         children: [
+          PageOneContainer(),
           Positioned(
             bottom: MediaQuery.of(context).size.height * 0.3,
             child: const PositionedColumn(),
           ),
         ],
+      ),
+    );
+  }
+}
+
+class PageOneContainer extends ConsumerWidget {
+  const PageOneContainer({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final isDarkMode = ref.watch(settingsNotifierProvider).isDarkMode;
+    const List<Color> gradientDark = [
+      Color(0xFF0D3A5C),
+      Color(0xFF306281),
+      Color(0xFFA2E6FA),
+    ];
+    const List<Color> gradientLight = [
+      Color(0xFF9F8FFF),
+      Color(0xFFA1D6FB),
+      Color(0xFFA2E6FA),
+    ];
+    return Container(
+      width: double.infinity,
+      height: double.infinity,
+      decoration: BoxDecoration(
+        gradient: LinearGradient(
+          colors: isDarkMode ? gradientDark : gradientLight,
+          begin: Alignment.topCenter,
+          end: Alignment.bottomCenter,
+          stops: const [0.0, 0.5, 1.0],
+        ),
       ),
     );
   }
@@ -53,67 +89,8 @@ class PositionedColumn extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
-          PageSelect(isDarkMode: true),
+          PageSelect(isDarkMode: false),
         ],
-      ),
-    );
-  }
-}
-
-class PageSelect extends ConsumerWidget {
-  const PageSelect({
-    super.key,
-    required this.isDarkMode,
-  });
-  final bool isDarkMode;
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return SizedBox(
-      width: MediaQuery.of(context).size.width * 0.6,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          ItemSelect(isSelected: true, isDarkMode: isDarkMode),
-          ItemSelect(isSelected: false, isDarkMode: isDarkMode),
-          ItemSelect(isSelected: false, isDarkMode: isDarkMode),
-          ItemSelect(isSelected: false, isDarkMode: isDarkMode),
-          ItemSelect(isSelected: false, isDarkMode: isDarkMode),
-        ],
-      ),
-    );
-  }
-}
-
-class ItemSelect extends StatelessWidget {
-  const ItemSelect({
-    super.key,
-    required this.isSelected,
-    required this.isDarkMode,
-  });
-  final bool isSelected;
-  final bool isDarkMode;
-  final int itemDarkSelect = 0xffA2E6FA;
-  final int itemLightSelect = 0xff0D3A5C;
-  final int itemDarkNotSelect = 0xff141414;
-  final int itemLightNotSelect = 0xffE6E6E6;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 28,
-      height: 7,
-      decoration: BoxDecoration(
-        borderRadius: const BorderRadius.all(
-          Radius.circular(20),
-        ),
-        color: isDarkMode
-            ? isSelected
-                ? Color(itemDarkSelect)
-                : Color(itemDarkNotSelect)
-            : isSelected
-                ? Color(itemLightSelect)
-                : Color(itemLightNotSelect),
       ),
     );
   }
